promstatsd: sane timestamps when no stats files yet
