+++ lib/util.h
@@ -238,6 +238,7 @@ extern double timeval_get_double(const struct timeval *tv);
 extern void timeval_set_double(struct timeval *tv, double d);
 extern void timeval_add_double(struct timeval *tv, double delta);
 extern double timesub(const struct timeval *start, const struct timeval *end);
+extern int64_t now_ms(void);
 
 extern clock_t sclock(void);
